windows/requirements-windows.txt
pystray>=0.19
Pillow>=10.0
qrcode>=7.0
pyttsx3>=2.90
requests>=2.28
